Started implementation of the free curve drawing tool
Implemented this tool by using a QPainterPath, and then creating a
QGraphicsPathItem. The problem is once the path is assigned to
QGraphicsPathItem, the later is not automatically updated, so the
assignment has to be done when the path is complete. This means that
when you are drawing, you see nothing until you finish. To solve this,
it may be needed to draw some temporary lines on the path and then
delete them when the QPainterPath is ready.