Finished first implementation of freehand drawing.
After struggling with updating QGraphicsPath with the newly added
points, the stategy was changed. A series of graphicsLine items are
created on the way while you are drawing the line. When it is finished
all these line items are removed and the painter path is put instead.
Different pens are used for the two stages. Still need to implement
the autopositioning of the path when resizing the window.