Implemented freeHand line adjusting when redrawing
Added the funcionality in drawElements to adjust the position of the
points that make up the painter path. For that, the graphisPathItem
stores a QVariant which is a QList<QVariant> that holds the points in
the Chart coordinates (these don't change when zooming or scrolling) and
these points are transformed into scene coordinates when redrawing. The
QGraphicsPath item has its internal path updated with this new path.