Began adding pen styles choices
Put a QFrame on the top right corner of the QGraphicsView to have the
line style controls. Tried both with and without an intermediate
QHBoxLayout. The image (QLabel) of the brush shows alright, but the
QCombo box for setting line width is not automatically placed besides
the brush image. When tried to use a layout, the image lost the desired
size and went back to its normal gigantic size. Still have to figure out
what is going on.